requirements.txt
Flask
python-dotenv
firebase-admin
google-cloud-storage